Implement role functionality
Implement the functionality to update any source to HTTPS for Debian
and any non-official source to HTTPS for Ubuntu. We must only update
non-official sources for Ubuntu because the official sources do not
support HTTPS at this time. Also add testing to verify that the
expected changes have been made.