Move the list of apt source files from defaults to vars
Contrary to appearances, this doesn't limit functionality.  This is
because, although the variable was previously in defaults, it was not
documented as a role variable in the README.md.

This also allows us to specify a different list of apt source files
for Debian Bullseye.  This is useful, since as of Debian Bullseye the
/etc/apt/sources.list file has moved to
/etc/apt/sources.list.d/debian.sources.